+++ packages/open-payments/src/client/index.test.ts
@@ -1,0 +1,91 @@
+import { createAuthenticatedClient } from '.'
+import fs from 'fs'
+import assert from 'assert'
+import { generateKeyPairSync } from 'crypto'
+import { silentLogger } from '../test/helpers'
+
+describe('Client', (): void => {
+  const TMP_DIR = './tmp'
+
+  beforeEach(async (): Promise<void> => {
+    fs.rmSync(TMP_DIR, { recursive: true, force: true })
+  })
+
+  afterEach(async (): Promise<void> => {
+    fs.rmSync(TMP_DIR, { recursive: true, force: true })
+  })
+
+  describe('createAuthenticatedClient', (): void => {
+    test('properly loads key with privateKey as file path', async (): Promise<void> => {
+      const keypair = generateKeyPairSync('ed25519')
+      const keyFilePath = `${TMP_DIR}/test-private-key.pem`
+      fs.mkdirSync(TMP_DIR)
+      fs.writeFileSync(
+        keyFilePath,
+        keypair.privateKey.export({ format: 'pem', type: 'pkcs8' })
+      )
+      assert.ok(fs.existsSync(keyFilePath))
+
+      await expect(
+        createAuthenticatedClient({
+          keyId: 'keyid-1',
+          walletAddressUrl: 'http://localhost:1000/.well-known/pay',
+          privateKey: keyFilePath,
+          logger: silentLogger
+        })
+      ).resolves.toBeDefined()
+    })
+
+    test('properly loads key with privateKey as KeyObject', async (): Promise<void> => {
+      const keypair = generateKeyPairSync('ed25519')
+
+      await expect(
+        createAuthenticatedClient({
+          logger: silentLogger,
+          keyId: 'keyid-1',
+          walletAddressUrl: 'http://localhost:1000/.well-known/pay',
+          privateKey: keypair.privateKey
+        })
+      ).resolves.toBeDefined()
+    })
+
+    test('properly loads key with privateKey as Buffer', async (): Promise<void> => {
+      const keypair = generateKeyPairSync('ed25519')
+
+      await expect(
+        createAuthenticatedClient({
+          logger: silentLogger,
+          keyId: 'keyid-1',
+          walletAddressUrl: 'http://localhost:1000/.well-known/pay',
+          privateKey: Buffer.from(
+            keypair.privateKey.export({ format: 'pem', type: 'pkcs8' })
+          )
+        })
+      ).resolves.toBeDefined()
+    })
+
+    test('throws error if could not load private key as Buffer', async (): Promise<void> => {
+      await expect(() =>
+        createAuthenticatedClient({
+          logger: silentLogger,
+          keyId: 'keyid-1',
+          walletAddressUrl: 'http://localhost:1000/.well-known/pay',
+          privateKey: Buffer.from('')
+        })
+      ).rejects.toThrow('Could not load private key. Key is not a valid file')
+    })
+
+    test('throws error if could not load private key', async (): Promise<void> => {
+      await expect(() =>
+        createAuthenticatedClient({
+          logger: silentLogger,
+          keyId: 'keyid-1',
+          walletAddressUrl: 'http://localhost:1000/.well-known/pay',
+          privateKey: '/incorrect/path/'
+        })
+      ).rejects.toThrow(
+        'Could not load private key. Key is not a valid path or file'
+      )
+    })
+  })
+})